Add remove / add/ and create
